Move ListViewGroupCollection to its own file without cleanup
